MainFormの結合クラスを減少。
